intensity dataset class and tests
